Refactoring: move some objects to 'utils' module
