+++ openprocurement/auction/tests/unit/conftest.py
@@ -1,30 +1,28 @@
-import couchdb
-import os
-import pytest
-import yaml
-from openprocurement.auction.chronograph import AuctionsChronograph
-from openprocurement.auction.tests.utils import update_auctionPeriod, \
-    AUCTION_DATA, PWD
-from openprocurement.auction.tests.unit.utils import TestClient
-from openprocurement.auction.tests.unit.utils import kill_child_processes
-from openprocurement.auction.worker.auction import Auction
-from gevent import spawn, killall, GreenletExit
 import json
 import logging
-from openprocurement.auction.helpers.chronograph import \
-    MAX_AUCTION_START_TIME_RESERV
+import couchdb
 import datetime
 import gc
-from greenlet import greenlet
-from openprocurement.auction.databridge import AuctionsDataBridge
-from copy import deepcopy
-from openprocurement.auction import core as core_module
 import openprocurement.auction.databridge as databridge_module
+import pytest
+from gevent import spawn, killall, GreenletExit
+from greenlet import greenlet
+from openprocurement.auction import core as core_module
+from openprocurement.auction.chronograph import AuctionsChronograph
+from openprocurement.auction.databridge import AuctionsDataBridge
+from openprocurement.auction.helpers.chronograph import \
+    MAX_AUCTION_START_TIME_RESERV
 from openprocurement.auction.tests.unit.utils import get_tenders_dummy
+from openprocurement.auction.tests.unit.utils import kill_child_processes
+from openprocurement.auction.worker.auction import Auction
+from openprocurement.auction.tests.utils import update_auctionPeriod, \
+    AUCTION_DATA
+from openprocurement.auction.tests.unit.utils import worker_defaults, \
+    test_chronograph_config, worker_defaults_file_path, test_bridge_config
+import yaml
 
 
 LOGGER = logging.getLogger('Log For Tests')
-CONF_FILES_FOLDER = os.path.join(PWD, "unit", "data")
 
 test_log_config = {
      'version': 1,
@@ -44,39 +42,6 @@ logging.config.dictConfig(test_log_config)
 
 auction_data_simple = AUCTION_DATA['simple']
 auction_data_multilot = AUCTION_DATA['multilot']
-
-worker_defaults_file_path = \
-    os.path.join(CONF_FILES_FOLDER, "auction_worker_defaults.yaml")
-with open(worker_defaults_file_path) as stream:
-    worker_defaults = yaml.load(stream)
-
-chronograph_conf_file_path = \
-    os.path.join(CONF_FILES_FOLDER, 'auctions_chronograph.yaml')
-with open(chronograph_conf_file_path) as stream:
-    test_chronograph_config = yaml.load(stream)
-    test_chronograph_config['disable_existing_loggers'] = False
-    test_chronograph_config['handlers']['journal']['formatter'] = 'simple'
-    test_chronograph_config['main']['auction_worker'] = \
-        os.path.join(PWD, (".." + os.path.sep)*5, "bin", "auction_worker")
-    test_chronograph_config['main']['auction_worker_config'] = \
-        os.path.join(PWD, 'unit', 'data', 'auction_worker_defaults.yaml')
-    test_chronograph_config['main'] \
-    ['auction_worker_config_for_api_version_dev'] = \
-        os.path.join(PWD, 'unit', 'data', 'auction_worker_defaults.yaml')
-
-databridge_conf_file_path = \
-    os.path.join(CONF_FILES_FOLDER, 'auctions_data_bridge.yaml')
-with open(databridge_conf_file_path) as stream:
-    test_bridge_config = yaml.load(stream)
-    test_bridge_config['disable_existing_loggers'] = False
-    test_bridge_config['handlers']['journal']['formatter'] = 'simple'
-
-test_bridge_config_error_port = deepcopy(test_bridge_config)
-couch_url = test_bridge_config_error_port['main']['couch_url']
-error_port = str(int(couch_url.split(':')[-1][:-1]) + 1)
-couch_url_parts = couch_url.split(':')[0:-1]
-couch_url_parts.append(error_port)
-test_bridge_config_error_port['main']['couch_url'] = ':'.join(couch_url_parts)
 
 
 @pytest.fixture(scope='function')
@@ -101,32 +66,6 @@ def db(request):
     request.addfinalizer(delete)
 
     return data_base
-
-
-# TODO: change host
-test_client = \
-    TestClient('http://0.0.0.0:{port}'.
-               format(port=test_chronograph_config['main'].get('web_app')))
-
-
-def job_is_added():
-    resp = test_client.get('jobs')
-    return (len(json.loads(resp.content)) == 1)
-
-
-def job_is_not_added():
-    resp = test_client.get('jobs')
-    return (len(json.loads(resp.content)) == 0)
-
-
-def job_is_active():
-    resp = test_client.get('active_jobs')
-    return (len(json.loads(resp.content)) == 1)
-
-
-def job_is_not_active():
-    resp = test_client.get('active_jobs')
-    return (len(json.loads(resp.content)) == 0)
 
 
 @pytest.fixture(scope='function')
@@ -164,8 +103,6 @@ def chronograph(request):
 
 @pytest.yield_fixture(scope="function")
 def auction(request):
-    # TODO: change it (MAX_AUCTION_START_TIME_RESERV)
-
     defaults = {'time': MAX_AUCTION_START_TIME_RESERV,
                 'delta_t': datetime.timedelta(seconds=10)}
