+++ openprocurement/auction/tests/unit/utils.py
@@ -1,13 +1,52 @@
 import contextlib
+import json
+
 from requests import Session as Sess
 import signal, psutil
 import os
 from copy import deepcopy
+from openprocurement.auction.tests.utils import PWD
+import yaml
 
 
 ID = 'UA-11111'
 LOT_ID = '11111111111111111111111111111111'
+API_EXTRA = {'opt_fields': 'status,auctionPeriod,lots,procurementMethodType', 'mode': '_all_'}
+CONF_FILES_FOLDER = os.path.join(PWD, "unit", "data")
 
+
+worker_defaults_file_path = \
+    os.path.join(CONF_FILES_FOLDER, "auction_worker_defaults.yaml")
+with open(worker_defaults_file_path) as stream:
+    worker_defaults = yaml.load(stream)
+
+chronograph_conf_file_path = \
+    os.path.join(CONF_FILES_FOLDER, 'auctions_chronograph.yaml')
+with open(chronograph_conf_file_path) as stream:
+    test_chronograph_config = yaml.load(stream)
+    test_chronograph_config['disable_existing_loggers'] = False
+    test_chronograph_config['handlers']['journal']['formatter'] = 'simple'
+    test_chronograph_config['main']['auction_worker'] = \
+        os.path.join(PWD, (".." + os.path.sep)*5, "bin", "auction_worker")
+    test_chronograph_config['main']['auction_worker_config'] = \
+        os.path.join(PWD, 'unit', 'data', 'auction_worker_defaults.yaml')
+    test_chronograph_config['main'] \
+    ['auction_worker_config_for_api_version_dev'] = \
+        os.path.join(PWD, 'unit', 'data', 'auction_worker_defaults.yaml')
+
+databridge_conf_file_path = \
+    os.path.join(CONF_FILES_FOLDER, 'auctions_data_bridge.yaml')
+with open(databridge_conf_file_path) as stream:
+    test_bridge_config = yaml.load(stream)
+    test_bridge_config['disable_existing_loggers'] = False
+    test_bridge_config['handlers']['journal']['formatter'] = 'simple'
+
+test_bridge_config_error_port = deepcopy(test_bridge_config)
+couch_url = test_bridge_config_error_port['main']['couch_url']
+error_port = str(int(couch_url.split(':')[-1][:-1]) + 1)
+couch_url_parts = couch_url.split(':')[0:-1]
+couch_url_parts.append(error_port)
+test_bridge_config_error_port['main']['couch_url'] = ':'.join(couch_url_parts)
 
 @contextlib.contextmanager
 def put_test_doc(db, doc):
@@ -35,7 +74,6 @@ def kill_child_processes(parent_pid=os.getpid(), sig=signal.SIGTERM):
     for process in children:
         process.send_signal(sig)
 
-# = = = = = = = = = Data for different statuses = = = = = = = = =
 # Data for test with 'active.auction' status
 tender_data_templ = {'id': ID, 'status': 'active.auction'}
 tender_data_wrong_status = {'id': ID, 'status': 'wrong.status'}
@@ -134,4 +172,27 @@ def get_tenders_dummy(tender_data_list):
     return a
 
 
-API_EXTRA = {'opt_fields': 'status,auctionPeriod,lots,procurementMethodType', 'mode': '_all_'}
+# TODO: change host
+test_client = \
+    TestClient('http://0.0.0.0:{port}'.
+               format(port=test_chronograph_config['main'].get('web_app')))
+
+
+def job_is_added():
+    resp = test_client.get('jobs')
+    return len(json.loads(resp.content)) == 1
+
+
+def job_is_not_added():
+    resp = test_client.get('jobs')
+    return len(json.loads(resp.content)) == 0
+
+
+def job_is_active():
+    resp = test_client.get('active_jobs')
+    return len(json.loads(resp.content)) == 1
+
+
+def job_is_not_active():
+    resp = test_client.get('active_jobs')
+    return len(json.loads(resp.content)) == 0
